refactor modal profileTag
